BE/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==4.2.14
django-allauth==0.63.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-mysql==4.14.0
djangorestframework==3.15.2
Faker==26.0.0
idna==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
